linalg-traits -- refactor files and ref_impl function names
