+++ rstsr-linalg-traits/Cargo.toml
@@ -10,5 +10,6 @@ license.workspace = true
 
 [dependencies]
 rstsr-core = { workspace = true, default-features = false, features = ["rayon"] }
+rstsr-common = { workspace = true, default-features = false, features = ["rayon"] }
 rstsr-blas-traits = { workspace = true, default-features = false }
 derive_builder = { workspace = true }
