+++ rstsr-linalg-traits/src/lib.rs
@@ -4,4 +4,4 @@
 pub mod prelude;
 pub mod prelude_dev;
 pub mod ref_impl_blas;
-pub mod traits;
+pub mod traits_def;
