+++ rstsr-linalg-traits/src/prelude_dev.rs
@@ -1,13 +1,2 @@
-pub use crate::traits::cholesky::*;
-pub use crate::traits::eigh::*;
-pub use crate::traits::inv::*;
-pub use crate::traits::solve_general::*;
-pub use crate::traits::solve_symmetric::*;
-pub use crate::traits::solve_triangular::*;
-
-pub use crate::ref_impl_blas::cholesky::*;
-pub use crate::ref_impl_blas::eigh::*;
-pub use crate::ref_impl_blas::inv::*;
-pub use crate::ref_impl_blas::solve_general::*;
-pub use crate::ref_impl_blas::solve_symmetric::*;
-pub use crate::ref_impl_blas::solve_triangular::*;
+pub use crate::ref_impl_blas::*;
+pub use crate::traits_def::*;
