--- rstsr-linalg-traits/src/ref_impl_blas/mod.rs
@@ -1,13 +1,0 @@
-pub mod cholesky;
-pub mod eigh;
-pub mod inv;
-pub mod solve_general;
-pub mod solve_symmetric;
-pub mod solve_triangular;
-
-pub use cholesky::*;
-pub use eigh::*;
-pub use inv::*;
-pub use solve_general::*;
-pub use solve_symmetric::*;
-pub use solve_triangular::*;
