--- rstsr-linalg-traits/src/ref_impl_blas/solve_general.rs
@@ -1,17 +1,0 @@
-use rstsr_blas_traits::prelude::*;
-use rstsr_core::prelude_dev::*;
-
-pub fn blas_solve_general_f<'b, T, B>(
-    a: TensorReference<'_, T, B, Ix2>,
-    b: TensorReference<'b, T, B, Ix2>,
-) -> Result<TensorMutable<'b, T, B, Ix2>>
-where
-    T: BlasFloat,
-    B: LapackDriverAPI<T>,
-{
-    let device = a.device().clone();
-    let nthreads = device.get_current_pool().map_or(1, |pool| pool.current_num_threads());
-    let task = || GESV::default().a(a).b(b).build()?.run();
-    let result = device.with_blas_num_threads(nthreads, task)?;
-    Ok(result.2)
-}
