--- rstsr-linalg-traits/src/traits/mod.rs
@@ -1,6 +1,0 @@
-pub mod cholesky;
-pub mod eigh;
-pub mod inv;
-pub mod solve_general;
-pub mod solve_symmetric;
-pub mod solve_triangular;
