--- rstsr-linalg-traits/src/traits/solve_triangular.rs
@@ -1,26 +1,0 @@
-use rstsr_core::prelude::rt::Result;
-
-pub trait SolveTriangularAPI<Inp> {
-    type Out;
-    fn solve_triangular_f(self) -> Result<Self::Out>;
-    fn solve_triangular(self) -> Self::Out
-    where
-        Self: Sized,
-    {
-        Self::solve_triangular_f(self).unwrap()
-    }
-}
-
-pub fn solve_triangular_f<Args, Inp>(args: Args) -> Result<<Args as SolveTriangularAPI<Inp>>::Out>
-where
-    Args: SolveTriangularAPI<Inp>,
-{
-    Args::solve_triangular_f(args)
-}
-
-pub fn solve_triangular<Args, Inp>(args: Args) -> <Args as SolveTriangularAPI<Inp>>::Out
-where
-    Args: SolveTriangularAPI<Inp>,
-{
-    Args::solve_triangular(args)
-}
